requirements.txt
asgiref==3.5.0
aspy.refactor-imports==2.2.1
asttokens==2.0.5
autoflake==1.4
black==22.1.0
cached-property==1.5.2
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.0.4
distlib==0.3.4
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
executing==0.8.1
filelock==3.6.0
flake8==4.0.1
identify==2.4.10
idna==3.3
Jinja2==3.0.3
mailjet-rest==1.3.4
MarkupSafe==2.1.0
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
pre-commit==2.17.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
pyupgrade==2.31.0
PyYAML==6.0
reorder-python-imports==2.7.1
requests==2.27.1
simple-print==1.5.1
six==1.16.0
sqlparse==0.4.2
termcolor==1.1.0
tokenize-rt==4.2.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
ujson==5.1.0
urllib3==1.26.8
virtualenv==20.13.1
yesqa==1.3.0
